requirements.txt
APScheduler==3.6.3
astroid==2.4.2
autopep8==1.5.4
Brotli==1.0.9
CacheControl==0.12.6
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
decorator==4.4.2
firebase-admin==4.4.0
flake8==3.8.4
Flask==1.1.2
Flask-Cache==0.13.1
Flask-Compress==1.8.0
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.0.0
gevent==20.9.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.33.2
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
passlib==1.7.4
Pillow==8.1.1
plivo==4.15.1
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-engineio==3.13.2
python-http-client==3.3.1
python-socketio==4.6.0
pytz==2020.4
requests==2.25.0
rsa==4.6
sendgrid==6.4.7
six==1.15.0
starkbank-ecdsa==1.1.0
stripe==2.55.0
toml==0.10.2
typed-ast==1.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
